Make top 100 Schools not show bottom 100 schools
This is a school outside Korogwe in NorthEastern Tanzania, one of the lowest performing districts according to the open data overall. It seems for secondary schools, the criteria for "Top School" and "Top 100 most improved" badge got flipped, same for the "Top 40" list itself. Nice job overall. To even be able to start asking these questions visually is amazing.

Before:
https://www.dropbox.com/s/i6lel5tqoq4wpxc/Screenshot%202015-10-20%2011.24.20.png?dl=0

After:
https://www.dropbox.com/s/5mko64dgstphm1e/Screenshot%202015-10-20%2011.24.07.png?dl=0